requirements.txt
requests>=2.31.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
aiohttp>=3.9.0
selenium>=4.15.0
duckduckgo-search>=6.0.0
